requirements.txt
pandas>=2.0
numpy>=1.24
openai>=1.0
python-dotenv>=1.0
openpyxl>=3.0
matplotlib>=3.7
markdown>=3.4
pdfkit>=1.0
xhtml2pdf>=0.2
